refactor(#741): refactor documentDuplicateController
