+++ api/controllers/v1/document-duplicate/delete-one.js
@@ -1,0 +1,39 @@
+const RightService = require('../../../services/RightService');
+
+module.exports = async (req, res) => {
+  const hasRight = await sails.helpers.checkRight
+    .with({
+      groups: req.token.groups,
+      rightEntity: RightService.RightEntities.DOCUMENT_DUPLICATE,
+      rightAction: RightService.RightActions.DELETE_ANY,
+    })
+    .intercept('rightNotFound', () =>
+      res.serverError(
+        'A server error occured when checking your right to delete an document duplicate.'
+      )
+    );
+  if (!hasRight) {
+    return res.forbidden(
+      'You are not authorized to delete a document duplicate.'
+    );
+  }
+  const id = req.param('id');
+  if (!id) {
+    return res.badRequest('You must provide the id of the duplicate.');
+  }
+
+  if (
+    !(await sails.helpers.checkIfExists.with({
+      attributeName: 'id',
+      attributeValue: req.param('id'),
+      sailsModel: TDocumentDuplicate,
+    }))
+  ) {
+    return res.badRequest(
+      `Could not find duplicate with id ${req.param('id')}.`
+    );
+  }
+
+  await TDocumentDuplicate.destroyOne(id);
+  return res.ok();
+};
